Fix return types of functions in src/wbuf.h

+++ src/wbuf.h
@@ -1,7 +1,7 @@
 #ifndef WBUF_H
 #define WBUF_H
 
-#include <stddef.h>
+#include <sys/types.h>
 
 typedef struct {
     size_t fill;
@@ -14,9 +14,9 @@ typedef struct {
     .fd = -1 \
 }
 
-int wbuf_flush(WriteBuffer *wbuf);
-int wbuf_write(WriteBuffer *wbuf, const char *buf, size_t count);
-int wbuf_write_str(WriteBuffer *wbuf, const char *str);
-int wbuf_write_ch(WriteBuffer *wbuf, char ch);
+ssize_t wbuf_flush(WriteBuffer *wbuf);
+ssize_t wbuf_write(WriteBuffer *wbuf, const char *buf, size_t count);
+ssize_t wbuf_write_str(WriteBuffer *wbuf, const char *str);
+ssize_t wbuf_write_ch(WriteBuffer *wbuf, char ch);
 
 #endif
